+++ src/backend/mailbox_manager/api/permissions.py
@@ -1,5 +1,7 @@
 """Permission handlers for the People mailbox manager app."""
 
+from rest_framework import permissions
+
 from core.api import permissions as core_permissions
 
 from mailbox_manager import models
@@ -14,7 +16,7 @@ class AccessPermission(core_permissions.IsAuthenticated):
         return abilities.get(request.method.lower(), False)
 
 
-class MailBoxPermission(core_permissions.IsAuthenticated):
+class MailBoxPermission(AccessPermission):
     """Permission class to manage mailboxes for a mail domain"""
 
     def has_permission(self, request, view):
@@ -22,6 +24,20 @@ class MailBoxPermission(core_permissions.IsAuthenticated):
         domain = models.MailDomain.objects.get(slug=view.kwargs.get("domain_slug", ""))
         abilities = domain.get_abilities(request.user)
         return abilities.get(request.method.lower(), False)
+
+
+class IsMailboxOwnerPermission(permissions.BasePermission):
+    """Authorize update for domain viewers on their own mailbox."""
+
+    def has_permission(self, request, view):
+        """This permission is specifically about updates"""
+        domain = models.MailDomain.objects.get(slug=view.kwargs.get("domain_slug", ""))
+        abilities = domain.get_abilities(request.user)
+        return abilities["get"]
+
+    def has_object_permission(self, request, view, obj):
+        """If the user is trying to update their own mailbox."""
+        return obj.get_email() == request.user.email
 
 
 class MailDomainAccessRolePermission(core_permissions.IsAuthenticated):